Refactored security models and its usage
