+++ fastapi/dependencies/utils.py
@@ -27,7 +27,7 @@ from fastapi.concurrency import (
 )
 from fastapi.dependencies.models import Dependant, SecurityRequirement
 from fastapi.logger import logger
-from fastapi.security.base import SecurityBase
+from fastapi.security.models import SecurityBase
 from fastapi.security.oauth2 import OAuth2, SecurityScopes
 from fastapi.security.open_id_connect_url import OpenIdConnect
 from fastapi.utils import create_response_field, get_path_param_names
