+++ fastapi/security/api_key.py
@@ -1,14 +1,10 @@
 from typing import Optional
 
 from fastapi.openapi.models import APIKey, APIKeyIn
-from fastapi.security.base import SecurityBase
+from fastapi.security.models import APIKeyBase
 from starlette.exceptions import HTTPException
 from starlette.requests import Request
 from starlette.status import HTTP_403_FORBIDDEN
-
-
-class APIKeyBase(SecurityBase):
-    pass
 
 
 class APIKeyQuery(APIKeyBase):
