--- fastapi/security/base.py
@@ -1,6 +1,0 @@
-from fastapi.openapi.models import SecurityBase as SecurityBaseModel
-
-
-class SecurityBase:
-    model: SecurityBaseModel
-    scheme_name: str
