+++ fastapi/security/open_id_connect_url.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
 from fastapi.openapi.models import OpenIdConnect as OpenIdConnectModel
-from fastapi.security.base import SecurityBase
+from fastapi.security.models import SecurityBase
 from starlette.exceptions import HTTPException
 from starlette.requests import Request
 from starlette.status import HTTP_403_FORBIDDEN
